requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
djangorestframework==3.15.2
ffmpeg==1.4
filelock==3.18.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib_resources==6.4.5
intasend-python==1.0.8
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.2
kiwisolver==1.4.7
MarkupSafe==2.1.3
matplotlib==3.9.2
mediapipe==0.10.10
ml_dtypes==0.5.0
multidict==6.1.0
neat-python==0.9
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
pillow==11.2.1
pipenv==2024.4.1
platformdirs==4.3.7
propcache==0.2.1
protobuf==3.20.3
pycountry==24.6.1
pycparser==2.22
pyFirmata==1.1.0
pygame==2.6.0
PyMySQL==1.1.1
pyparsing==3.2.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.4.3.2
PyQt5_sip==12.15.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytube==15.0.0
qasync==0.27.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
scipy==1.13.1
six==1.16.0
sounddevice==0.5.1
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.29.3
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.18.3
zipp==3.15.0
